add and implement method add() of the CommonDAO interface

+++ core/src/main/java/com/z_soft/z_finance/core/impls/dao/StorageDAOImpl.java
@@ -1,15 +1,18 @@
 package com.z_soft.z_finance.core.impls.dao;
 
 import com.z_soft.z_finance.core.database.SQLiteConnection;
+import com.z_soft.z_finance.core.exceptions.CurrencyException;
 import com.z_soft.z_finance.core.impls.DefaultStorage;
 import com.z_soft.z_finance.core.interfaces.Storage;
 import com.z_soft.z_finance.core.interfaces.dao.StorageDAO;
 
 import java.math.BigDecimal;
+import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.sql.Types;
 import java.util.ArrayList;
 import java.util.Currency;
 import java.util.List;
@@ -28,34 +31,57 @@ public class StorageDAOImpl implements StorageDAO {
     public List<Storage> getAll() {
         storageList.clear();
 
-        try (Statement stmt = SQLiteConnection.getConnection().createStatement();
-             ResultSet rs = stmt.executeQuery("select * from " + STORAGE_TABLE + " order by parent_id")) {
+        try {
 
-            while (rs.next()) {
-                DefaultStorage storage = new DefaultStorage();
-                storage.setId(rs.getLong("id"));
-                storage.setName(rs.getString("name"));
-                storage.setParentId(rs.getLong("parent_id"));
-                storageList.add(storage);
+            try (Statement stmt = SQLiteConnection.getConnection().createStatement();
+                 ResultSet rs = stmt.executeQuery("select * from " + STORAGE_TABLE + " order by parent_id")) {
+
+                while (rs.next()) {
+                    DefaultStorage storage = new DefaultStorage();
+                    storage.setId(rs.getLong("id"));
+                    storage.setName(rs.getString("name"));
+                    storage.setParentId(rs.getLong("parent_id"));
+
+                    storageList.add(storage);
+                }
+
+            }
+
+            // для каждого хранилища загрузить доступны валюты и баланс
+            for (Storage storage : storageList) {
+
+                try (PreparedStatement stmt = SQLiteConnection.getConnection().prepareStatement("select * from " + CURRENCY_AMOUNT_TABLE + " where storage_id =?")) {
+
+                    stmt.setLong(1, storage.getId());
+                    try(ResultSet rs = stmt.executeQuery()){
+                        while (rs.next()) {
+                            storage.addCurrency(Currency.getInstance(rs.getString("currency_code")), rs.getBigDecimal("amount"));
+                        }
+                    }
+                }
+
             }
 
             return storageList;
 
         } catch (SQLException e) {
             Logger.getLogger(StorageDAOImpl.class.getName()).log(Level.SEVERE, null, e);
+        } catch (CurrencyException e) {
+            e.printStackTrace();
         }
 
         return null;
     }
 
     @Override
-    public boolean addCurrency(Storage storage, Currency currency) {
+    public boolean addCurrency(Storage storage, Currency currency, BigDecimal initAmount) {
 
         // для автоматического закрытия ресурсов
-        try (PreparedStatement stmt = SQLiteConnection.getConnection().prepareStatement("insert into " + CURRENCY_AMOUNT_TABLE + "(currency_code, storage_id) values(?,?)")) {
+        try (PreparedStatement stmt = SQLiteConnection.getConnection().prepareStatement("insert into " + CURRENCY_AMOUNT_TABLE + "(currency_code, storage_id, amount) values(?,?,?)")) {
 
             stmt.setString(1, currency.getCurrencyCode());
             stmt.setLong(2, storage.getId());
+            stmt.setBigDecimal(3, initAmount);
 
             if (stmt.executeUpdate() == 1) { // если была добавлена 1 запись
                 return true;
@@ -71,24 +97,32 @@ public class StorageDAOImpl implements StorageDAO {
 
     @Override
     public boolean deleteCurrency(Storage storage, Currency currency) {
-        // TODO реализовать - если есть ли операции по данной валюте - запрещать удаление
-        try (PreparedStatement stmt = SQLiteConnection.getConnection().prepareStatement("delete from " + CURRENCY_AMOUNT_TABLE + " where storage_id=? and currency_code=?")) {
 
-            stmt.setLong(1, storage.getId());
-            stmt.setString(2, currency.getCurrencyCode());
+        if (!operationExist(storage, currency)) {// если есть операции, которые привязаны к этому хранилищцу и валюте
+
+            try (PreparedStatement stmt = SQLiteConnection.getConnection().prepareStatement("delete from " + CURRENCY_AMOUNT_TABLE + " where storage_id=? and currency_code=?")) {
+
+                stmt.setLong(1, storage.getId());
+                stmt.setString(2, currency.getCurrencyCode());
 
 
-            if (stmt.executeUpdate() == 1) {  // если была обновлена 1 запись
-                return true;
+                if (stmt.executeUpdate() == 1) {  // если была обновлена 1 запись
+                    return true;
+                }
+
+            } catch (SQLException e) {
+                Logger.getLogger(StorageDAOImpl.class.getName()).log(Level.SEVERE, null, e);
             }
 
-        } catch (SQLException e) {
-            Logger.getLogger(StorageDAOImpl.class.getName()).log(Level.SEVERE, null, e);
         }
 
         return false;
     }
 
+    private boolean operationExist(Storage storage, Currency currency) {
+        // TODO реализовать проверку, есть ли операции по этому хранилищу и валюте
+        return false;
+    }
     @Override
     public boolean updateAmount(Storage storage, Currency currency, BigDecimal amount) {
         try (PreparedStatement stmt = SQLiteConnection.getConnection().prepareStatement("update " + CURRENCY_AMOUNT_TABLE + " set amount=? where storage_id=? and currency_code=?")) {
@@ -168,6 +202,69 @@ public class StorageDAOImpl implements StorageDAO {
 
         } catch (SQLException e) {
             Logger.getLogger(StorageDAOImpl.class.getName()).log(Level.SEVERE, null, e);
+        }
+
+        return false;
+    }
+
+    @Override
+    public boolean add(Storage storage) {
+
+        // для хранилища нужно вставлять данные в разные таблицы, поэтому выполняем в одной транзакции
+        Connection con = SQLiteConnection.getConnection();
+
+        try {
+            con.setAutoCommit(false);// включаем режим ручного подтверждения транзакции (commit)
+
+
+            // само добавляем само хранилище, т.к. в таблице валют работает foreign key
+            try (PreparedStatement stmt = con.prepareStatement("insert into " + STORAGE_TABLE + "(name, parent_id) values(?,?)", Statement.RETURN_GENERATED_KEYS)) {// возвращать id вставленной записи
+
+                stmt.setString(1, storage.getName());
+
+                if (storage.hasParent()) {
+                    stmt.setLong(2, storage.getParent().getId());
+                } else {
+                    stmt.setNull(2, Types.BIGINT);
+                }
+
+
+                if (stmt.executeUpdate() == 1) {// если хранилище добавилось нормально
+                    try (ResultSet rs = stmt.getGeneratedKeys()) {// получаем id вставленной записи
+
+                        if (rs.next()) {
+                            storage.setId(rs.getLong(1));// не забываем просвоить новый id в объект, иначе дальше не добавятся валюты из-за foreign key
+                        }
+
+                        // вставляем все валюты с суммами
+                        for (Currency c : storage.getAvailableCurrencies()) {
+                            if (!addCurrency(storage, c, storage.getAmount(c))) {// если любой из запросов (добавление валюты) не выполнится - откатываем всю транзакцию
+                                con.rollback();
+                                return false;
+                            }
+
+                        }
+
+                        con.commit();// если все выполнилось без ошибок - подтверждаем транзакцию
+                        return true;
+                    }
+
+                }
+
+                con.rollback();// если код дошел до этого места - значит что-то пошлое не так, поэтому откатываем транзакцию
+
+            }
+
+        } catch (SQLException e) {
+            e.printStackTrace();
+        } catch (CurrencyException e) {
+            e.printStackTrace();
+        } finally {
+            try {
+                con.setAutoCommit(true);// возвращаем настройку обратно
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
         }
 
         return false;
